% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemblRestWrapper.R
\name{get_gene_id}
\alias{get_gene_id}
\title{Get gene name from protein ID}
\usage{
get_gene_id(id)
}
\arguments{
\item{id}{The protein id}
}
\description{
Retrieves a gene name from a provided protein ID using Uniprot.
}
