% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemblRestWrapper.R
\name{get_GO_info}
\alias{get_GO_info}
\title{Get GO information}
\usage{
get_GO_info(id)
}
\arguments{
\item{id}{The gene identifier you would like GO information for.}
}
\description{
Retrieves GO term info relating to a gene ID.
}
